Move dashboard cells into cells.go
